Move API handling to requests folder
